refactor: move fs deployments
