+++ packages/rust-serverless/src/deployments/filesystem.rs
@@ -1,0 +1,44 @@
+use std::io::Write;
+use std::{env, fs, io, path::Path};
+
+use super::Deployment;
+
+pub fn has_deployment_code(deployment: &Deployment) -> bool {
+    let path = Path::new("deployments").join(deployment.id.clone() + ".js");
+
+    path.exists()
+}
+
+pub fn get_deployment_code(deployment: &Deployment) -> io::Result<String> {
+    let path = Path::new(env::current_dir().unwrap().as_path())
+        .join("deployments")
+        .join(deployment.id.clone() + ".js");
+    let code = fs::read_to_string(path)?;
+
+    Ok(code)
+}
+
+pub fn write_deployment(deployment_id: String, buf: &[u8]) -> io::Result<()> {
+    let mut file = fs::File::create(Path::new("deployments").join(deployment_id + ".js"))?;
+
+    file.write_all(buf)?;
+
+    Ok(())
+}
+
+pub fn write_deployment_asset(deployment_id: String, asset: &str, buf: &[u8]) -> io::Result<()> {
+    let mut file = fs::File::create(Path::new("deployments").join(deployment_id + "/" + asset))?;
+
+    file.write_all(buf)?;
+
+    Ok(())
+}
+
+pub fn rm_deployment(deployment_id: String) -> io::Result<()> {
+    fs::remove_file(Path::new("deployments").join(deployment_id.clone() + ".js"))?;
+
+    // It's possible that folder doesn't exists
+    fs::remove_dir(Path::new("deployments").join(deployment_id)).unwrap_or(());
+
+    Ok(())
+}
